requirements.txt
psycopg2-binary
bcrypt
tabulate

